backend/requirements.txt
fastapi
uvicorn
Pillow
python-multipart